chore: add the `TODOs` file to gitignore
I usually keep `TODOs` file in the repository to keep track of the
highest priority issues that I'd like to work on next.

It isn't listed in a global gitignore configuration because that won't
be picked up by npm, so the file could still be accidentally published
to the npm registry, unless in projects that have the `files` field in
`package.json`. Therefore, I choose to manually ignore the `TODOs` file
in each project.